+++ src/wakatime.ts
@@ -27,9 +27,7 @@ export class WakaTime {
   };
   private agentName: string;
   private extension;
-  private statusBar: vscode.StatusBarItem = vscode.window.createStatusBarItem(
-    vscode.StatusBarAlignment.Left,
-  );
+  private statusBar?: vscode.StatusBarItem = undefined;
   private disposable: vscode.Disposable;
   private lastFile: string;
   private lastHeartbeat: number = 0;
@@ -37,109 +35,135 @@ export class WakaTime {
   private dependencies: Dependencies;
   private options: Options;
   private logger: Logger;
-  private getCodingActivityTimeout: number;
   private fetchTodayInterval: number = 60000;
+  private fetchTodayIntervalId?: NodeJS.Timeout;
   private lastFetchToday: number = 0;
   private showStatusBar: boolean;
   private showCodingActivity: boolean;
   private disabled: boolean = true;
   private extensionPath: string;
 
-  constructor(extensionPath: string, logger: Logger, options: Options) {
+  constructor(extensionPath: string, logger: Logger) {
     this.extensionPath = extensionPath;
     this.logger = logger;
-    this.options = options;
+    this.options = new Options(logger);
   }
 
-  public initialize(global: boolean): void {
-    this.dependencies = new Dependencies(this.options, this.logger, this.extensionPath, global);
-    this.statusBar.command = COMMAND_DASHBOARD;
+  public initialize(): void {
+    this.options.getSetting('settings', 'debug', false, (setting: Setting) => {
+      if (setting.value === 'true') {
+        this.logger.setLevel(LogLevel.DEBUG);
+      }
+      this.options.getSetting('settings', 'global', false, (setting: Setting) => {
+        const global = setting.value === 'true';
+        this.dependencies = new Dependencies(this.options, this.logger, this.extensionPath, global);
 
-    let extension = vscode.extensions.getExtension('WakaTime.vscode-wakatime');
-    this.extension = (extension != undefined && extension.packageJSON) || { version: '0.0.0' };
-    this.logger.debug(`Initializing WakaTime v${this.extension.version}`);
-    this.agentName = this.appNames[vscode.env.appName] || 'vscode';
-    this.statusBar.text = '$(clock) WakaTime Initializing...';
-    this.statusBar.show();
+        let extension = vscode.extensions.getExtension('WakaTime.vscode-wakatime');
+        this.extension = (extension != undefined && extension.packageJSON) || { version: '0.0.0' };
+        this.agentName = this.appNames[vscode.env.appName] || 'vscode';
 
-    this.setupEventListeners();
-
-    this.options.getSetting(
-      'settings',
-      'disabled',
-      this.options.getConfigFile(),
-      (disabled: Setting) => {
-        this.disabled = disabled.value === 'true';
-        if (this.disabled) {
-          this.setStatusBarVisibility(false);
-          this.logger.debug('Extension disabled, will not report coding stats to dashboard.');
-          return;
-        }
-
-        this.checkApiKey();
-        this.initializeDependencies();
-      },
-    );
-  }
-
-  public initializeDependencies(): void {
-    this.dependencies.checkAndInstall(() => {
-      this.logger.debug('WakaTime initialized.');
-      this.statusBar.text = '$(clock)';
-      this.statusBar.tooltip = 'WakaTime: Initialized';
-      this.options.getSetting(
-        'settings',
-        'status_bar_enabled',
-        this.options.getConfigFile(),
-        (setting: Setting) => {
-          this.showStatusBar = setting.value !== 'false';
-          this.setStatusBarVisibility(this.showStatusBar);
-        },
-      );
-      this.options.getSetting(
-        'settings',
-        'status_bar_coding_activity',
-        this.options.getConfigFile(),
-        (setting: Setting) => {
-          if (setting.value == 'false') {
-            this.showCodingActivity = false;
-          } else {
-            this.showCodingActivity = true;
-            this.getCodingActivity();
+        this.options.getSetting('settings', 'disabled', false, (disabled: Setting) => {
+          this.disabled = disabled.value === 'true';
+          if (this.disabled) {
+            this.dispose();
+            return;
           }
-        },
-      );
+
+          this.initializeDependencies();
+        });
+      });
     });
   }
 
-  public promptForApiKey(): void {
+  public dispose() {
+    this.clearTodayInterval();
+    this.statusBar?.dispose();
+    this.disposable.dispose();
+  }
+
+  public initializeDependencies(): void {
+    this.logger.debug(`Initializing WakaTime v${this.extension.version}`);
+
+    this.statusBar = vscode.window.createStatusBarItem(
+      vscode.StatusBarAlignment.Left,
+    );
+    this.statusBar.command = COMMAND_DASHBOARD;
+
     this.options.getSetting(
       'settings',
-      'api_key',
-      this.options.getConfigFile(),
-      (setting: Setting) => {
-        let defaultVal = setting.value;
-        if (Utils.validateKey(defaultVal) != '') defaultVal = '';
-        let promptOptions = {
-          prompt: 'WakaTime Api Key',
-          placeHolder: 'Enter your api key from https://wakatime.com/settings',
-          value: defaultVal,
-          ignoreFocusOut: true,
-          validateInput: Utils.validateKey.bind(this),
-        };
-        vscode.window.showInputBox(promptOptions).then(val => {
-          if (val != undefined) {
-            let validation = Utils.validateKey(val);
-            if (validation === '') this.options.setSetting('settings', 'api_key', val);
-            else vscode.window.setStatusBarMessage(validation);
-          } else vscode.window.setStatusBarMessage('WakaTime api key not provided');
-        });
+      'status_bar_enabled',
+      false,
+      (statusBarEnabled: Setting) => {
+        this.showStatusBar = statusBarEnabled.value !== 'false';
+        this.setStatusBarVisibility(this.showStatusBar);
+        this.updateStatusBarText('WakaTime Initializing...');
+
+        this.checkApiKey();
+
+        this.setupEventListeners();
+
+        this.options.getSetting(
+          'settings',
+          'status_bar_coding_activity',
+          false,
+          (showCodingActivity: Setting) => {
+            this.showCodingActivity = showCodingActivity.value !== 'false';
+
+            this.dependencies.checkAndInstall(() => {
+              this.logger.debug('WakaTime initialized');
+              this.updateStatusBarText();
+              this.updateStatusBarTooltip('WakaTime: Initialized');
+              this.getCodingActivity();
+            });
+
+          },
+        );
       },
     );
   }
 
+  private updateStatusBarText(text?: string): void {
+    if (!this.statusBar) return;
+    if (!text) {
+      this.statusBar.text = '$(clock)';
+    } else {
+      this.statusBar.text = '$(clock) ' +  text;
+    }
+  }
+
+  private updateStatusBarTooltip(tooltipText: string): void {
+    if (!this.statusBar) return;
+    this.statusBar.tooltip = tooltipText;
+  }
+
+  private statusBarShowingError(): boolean {
+    if (!this.statusBar) return false;
+    return this.statusBar.text.indexOf('Error') != -1;
+  }
+
+  public promptForApiKey(): void {
+    this.options.getSetting('settings', 'api_key', false, (setting: Setting) => {
+      let defaultVal = setting.value;
+      if (Utils.apiKeyInvalid(defaultVal)) defaultVal = '';
+      let promptOptions = {
+        prompt: 'WakaTime Api Key',
+        placeHolder: 'Enter your api key from https://wakatime.com/settings',
+        value: defaultVal,
+        ignoreFocusOut: true,
+        validateInput: Utils.apiKeyInvalid.bind(this),
+      };
+      vscode.window.showInputBox(promptOptions).then((val) => {
+        if (val != undefined) {
+          let invalid = Utils.apiKeyInvalid(val);
+          if (!invalid) this.options.setSetting('settings', 'api_key', val, false);
+          else vscode.window.setStatusBarMessage(invalid);
+        } else vscode.window.setStatusBarMessage('WakaTime api key not provided');
+      });
+    });
+  }
+
   public promptForProxy(): void {
-    this.options.getSetting('settings', 'proxy', this.options.getConfigFile(), (proxy: Setting) => {
+    this.options.getSetting('settings', 'proxy', false, (proxy: Setting) => {
       let defaultVal = proxy.value;
       if (!defaultVal) defaultVal = '';
       let promptOptions = {
@@ -149,14 +173,14 @@ export class WakaTime {
         ignoreFocusOut: true,
         validateInput: Utils.validateProxy.bind(this),
       };
-      vscode.window.showInputBox(promptOptions).then(val => {
-        if (val || val === '') this.options.setSetting('settings', 'proxy', val);
+      vscode.window.showInputBox(promptOptions).then((val) => {
+        if (val || val === '') this.options.setSetting('settings', 'proxy', val, false);
       });
     });
   }
 
   public promptForDebug(): void {
-    this.options.getSetting('settings', 'debug', this.options.getConfigFile(), (debug: Setting) => {
+    this.options.getSetting('settings', 'debug', false, (debug: Setting) => {
       let defaultVal = debug.value;
       if (!defaultVal || defaultVal !== 'true') defaultVal = 'false';
       let items: string[] = ['true', 'false'];
@@ -165,9 +189,9 @@ export class WakaTime {
         value: defaultVal,
         ignoreFocusOut: true,
       };
-      vscode.window.showQuickPick(items, promptOptions).then(newVal => {
+      vscode.window.showQuickPick(items, promptOptions).then((newVal) => {
         if (newVal == null) return;
-        this.options.setSetting('settings', 'debug', newVal);
+        this.options.setSetting('settings', 'debug', newVal, false);
         if (newVal === 'true') {
           this.logger.setLevel(LogLevel.DEBUG);
           this.logger.debug('Debug enabled');
@@ -179,93 +203,78 @@ export class WakaTime {
   }
 
   public promptToDisable(): void {
-    this.options.getSetting(
-      'settings',
-      'disabled',
-      this.options.getConfigFile(),
-      (setting: Setting) => {
-        let currentVal = setting.value;
-        if (!currentVal || currentVal !== 'true') currentVal = 'false';
-        let items: string[] = ['disable', 'enable'];
-        const helperText = currentVal === 'true' ? 'disabled' : 'enabled';
-        let promptOptions = {
-          placeHolder: `disable or enable (extension is currently "${helperText}")`,
-          ignoreFocusOut: true,
-        };
-        vscode.window.showQuickPick(items, promptOptions).then(newVal => {
-          if (newVal !== 'enable' && newVal !== 'disable') return;
-          this.disabled = newVal === 'disable';
+    this.options.getSetting('settings', 'disabled', false, (setting: Setting) => {
+      const previousValue = this.disabled;
+      let currentVal = setting.value;
+      if (!currentVal || currentVal !== 'true') currentVal = 'false';
+      let items: string[] = ['disable', 'enable'];
+      const helperText = currentVal === 'true' ? 'disabled' : 'enabled';
+      let promptOptions = {
+        placeHolder: `disable or enable (extension is currently "${helperText}")`,
+        ignoreFocusOut: true,
+      };
+      vscode.window.showQuickPick(items, promptOptions).then((newVal) => {
+        if (newVal !== 'enable' && newVal !== 'disable') return;
+        this.disabled = newVal === 'disable';
+        if (this.disabled != previousValue) {
           if (this.disabled) {
-            this.options.setSetting('settings', 'disabled', 'true');
-            this.setStatusBarVisibility(false);
-            this.logger.debug('Extension disabled, will not report coding stats to dashboard.');
+            this.options.setSetting('settings', 'disabled', 'true', false);
+            this.logger.debug('Extension disabled, will not report code stats to dashboard');
+            this.dispose();
           } else {
-            this.options.setSetting('settings', 'disabled', 'false');
-            this.checkApiKey();
+            this.options.setSetting('settings', 'disabled', 'false', false);
             this.initializeDependencies();
-            if (this.showStatusBar) this.setStatusBarVisibility(true);
-            this.logger.debug('Extension enabled and reporting coding stats to dashboard.');
           }
-        });
-      },
-    );
+        }
+      });
+    });
   }
 
   public promptStatusBarIcon(): void {
-    this.options.getSetting(
-      'settings',
-      'status_bar_enabled',
-      this.options.getConfigFile(),
-      (setting: Setting) => {
-        let defaultVal = setting.value;
-        if (!defaultVal || defaultVal !== 'false') defaultVal = 'true';
-        let items: string[] = ['true', 'false'];
-        let promptOptions = {
-          placeHolder: `true or false (current value \"${defaultVal}\")`,
-          value: defaultVal,
-          ignoreFocusOut: true,
-        };
-        vscode.window.showQuickPick(items, promptOptions).then(newVal => {
-          if (newVal !== 'true' && newVal !== 'false') return;
-          this.options.setSetting('settings', 'status_bar_enabled', newVal);
-          this.showStatusBar = newVal === 'true'; // cache setting to prevent reading from disc too often
-          this.setStatusBarVisibility(this.showStatusBar);
-        });
-      },
-    );
+    this.options.getSetting('settings', 'status_bar_enabled', false, (setting: Setting) => {
+      let defaultVal = setting.value;
+      if (!defaultVal || defaultVal !== 'false') defaultVal = 'true';
+      let items: string[] = ['true', 'false'];
+      let promptOptions = {
+        placeHolder: `true or false (current value \"${defaultVal}\")`,
+        value: defaultVal,
+        ignoreFocusOut: true,
+      };
+      vscode.window.showQuickPick(items, promptOptions).then((newVal) => {
+        if (newVal !== 'true' && newVal !== 'false') return;
+        this.options.setSetting('settings', 'status_bar_enabled', newVal, false);
+        this.showStatusBar = newVal === 'true'; // cache setting to prevent reading from disc too often
+        this.setStatusBarVisibility(this.showStatusBar);
+      });
+    });
   }
 
   public promptStatusBarCodingActivity(): void {
-    this.options.getSetting(
-      'settings',
-      'status_bar_coding_activity',
-      this.options.getConfigFile(),
-      (setting: Setting) => {
-        let defaultVal = setting.value;
-        if (!defaultVal || defaultVal !== 'false') defaultVal = 'true';
-        let items: string[] = ['true', 'false'];
-        let promptOptions = {
-          placeHolder: `true or false (current value \"${defaultVal}\")`,
-          value: defaultVal,
-          ignoreFocusOut: true,
-        };
-        vscode.window.showQuickPick(items, promptOptions).then(newVal => {
-          if (newVal !== 'true' && newVal !== 'false') return;
-          this.options.setSetting('settings', 'status_bar_coding_activity', newVal);
-          if (newVal === 'true') {
-            this.logger.debug('Coding activity in status bar has been enabled');
-            this.showCodingActivity = true;
-            this.getCodingActivity(true);
-          } else {
-            this.logger.debug('Coding activity in status bar has been disabled');
-            this.showCodingActivity = false;
-            if (this.statusBar.text.indexOf('Error') == -1) {
-              this.statusBar.text = '$(clock)';
-            }
+    this.options.getSetting('settings', 'status_bar_coding_activity', false, (setting: Setting) => {
+      let defaultVal = setting.value;
+      if (!defaultVal || defaultVal !== 'false') defaultVal = 'true';
+      let items: string[] = ['true', 'false'];
+      let promptOptions = {
+        placeHolder: `true or false (current value \"${defaultVal}\")`,
+        value: defaultVal,
+        ignoreFocusOut: true,
+      };
+      vscode.window.showQuickPick(items, promptOptions).then((newVal) => {
+        if (newVal !== 'true' && newVal !== 'false') return;
+        this.options.setSetting('settings', 'status_bar_coding_activity', newVal, false);
+        if (newVal === 'true') {
+          this.logger.debug('Coding activity in status bar has been enabled');
+          this.showCodingActivity = true;
+          this.getCodingActivity();
+        } else {
+          this.logger.debug('Coding activity in status bar has been disabled');
+          this.showCodingActivity = false;
+          if (!this.statusBarShowingError()) {
+            this.updateStatusBarText();
           }
-        });
-      },
-    );
+        }
+      });
+    });
   }
 
   public openDashboardWebsite(): void {
@@ -274,7 +283,7 @@ export class WakaTime {
   }
 
   public openConfigFile(): void {
-    let path = this.options.getConfigFile();
+    let path = this.options.getConfigFile(false);
     if (path) {
       let uri = vscode.Uri.file(path);
       vscode.window.showTextDocument(uri);
@@ -289,34 +298,20 @@ export class WakaTime {
     }
   }
 
-  public dispose() {
-    this.statusBar.dispose();
-    this.disposable.dispose();
-    clearTimeout(this.getCodingActivityTimeout);
-  }
-
   private checkApiKey(): void {
-    this.hasApiKey(hasApiKey => {
+    this.options.hasApiKey((hasApiKey) => {
       if (!hasApiKey) this.promptForApiKey();
     });
   }
 
-  private hasApiKey(callback: (arg0: boolean) => void): void {
-    this.options
-      .getApiKeyAsync()
-      .then(apiKey => callback(Utils.validateKey(apiKey) === ''))
-      .catch(err => {
-        this.logger.error(`Error reading api key: ${err}`);
-        callback(false);
-      });
-  }
-
   private setStatusBarVisibility(isVisible: boolean): void {
     if (isVisible) {
-      this.statusBar.show();
+      this.setTodayInterval();
+      this.statusBar?.show();
       this.logger.debug('Status bar icon enabled.');
     } else {
-      this.statusBar.hide();
+      this.clearTodayInterval();
+      this.statusBar?.hide();
       this.logger.debug('Status bar icon disabled.');
     }
   }
@@ -328,7 +323,7 @@ export class WakaTime {
     vscode.window.onDidChangeActiveTextEditor(this.onChange, this, subscriptions);
     vscode.workspace.onDidSaveTextDocument(this.onSave, this, subscriptions);
 
-    // create a combined disposable from both event subscriptions
+    // create a combined disposable for all event subscriptions
     this.disposable = vscode.Disposable.from(...subscriptions);
   }
 
@@ -367,9 +362,9 @@ export class WakaTime {
     lines: number,
     isWrite: boolean,
   ): void {
-    this.hasApiKey(hasApiKey => {
-      if (hasApiKey) {
-        this._sendHeartbeat(file, time, selection, lines, isWrite);
+    this.options.getApiKey((apiKey) => {
+      if (apiKey) {
+        this._sendHeartbeat(apiKey, file, time, selection, lines, isWrite);
       } else {
         this.promptForApiKey();
       }
@@ -377,6 +372,7 @@ export class WakaTime {
   }
 
   private _sendHeartbeat(
+    apiKey: string,
     file: string,
     time: number,
     selection: vscode.Position,
@@ -388,24 +384,38 @@ export class WakaTime {
     // prevent sending the same heartbeat (https://github.com/wakatime/vscode-wakatime/issues/163)
     if (isWrite && this.isDuplicateHeartbeat(file, time, selection)) return;
 
+    let args: string[] = [];
+
+    args.push('--entity', Utils.quote(file));
+
     let user_agent =
       this.agentName + '/' + vscode.version + ' vscode-wakatime/' + this.extension.version;
-    let args = ['--entity', Utils.quote(file), '--plugin', Utils.quote(user_agent)];
+    args.push('--plugin', Utils.quote(user_agent));
+
     args.push('--lineno', String(selection.line + 1));
     args.push('--cursorpos', String(selection.character + 1));
     args.push('--lines-in-file', String(lines));
+
     let project = this.getProjectName(file);
     if (project) args.push('--alternate-project', Utils.quote(project));
+
+    let folder = this.getProjectFolder(file);
+    if (folder) args.push('--project-folder', Utils.quote(folder));
+
     if (isWrite) args.push('--write');
-    if (process.env.WAKATIME_API_KEY) args.push('--key', Utils.quote(process.env.WAKATIME_API_KEY));
+
+    args.push('--key', Utils.quote(apiKey));
+
     if (Dependencies.isWindows() || Dependencies.isPortable()) {
       args.push(
         '--config',
-        Utils.quote(this.options.getConfigFile()),
+        Utils.quote(this.options.getConfigFile(false)),
         '--log-file',
         Utils.quote(this.options.getLogFile()),
       );
     }
+
+    if (Utils.isUnsavedFile(file)) args.push('--is-unsaved-entity')
 
     const binary = this.dependencies.getCliLocation();
     this.logger.debug(`Sending heartbeat: ${Utils.formatArguments(binary, args)}`);
@@ -419,69 +429,74 @@ export class WakaTime {
     });
     proc.on('close', (code, _signal) => {
       if (code == 0) {
-        if (this.showStatusBar) {
-          if (!this.showCodingActivity) this.statusBar.text = '$(clock)';
-          this.getCodingActivity();
-        }
+        if (this.showStatusBar) this.getCodingActivity();
         this.logger.debug(`last heartbeat sent ${Utils.formatDate(new Date())}`);
       } else if (code == 102) {
         if (this.showStatusBar) {
-          if (!this.showCodingActivity) this.statusBar.text = '$(clock)';
-          this.statusBar.tooltip =
-            'WakaTime: working offline... coding activity will sync next time we are online';
+          if (!this.showCodingActivity) this.updateStatusBarText();
+          this.updateStatusBarTooltip('WakaTime: working offline... coding activity will sync next time we are online');
         }
         this.logger.warn(
-          `Api eror (102); Check your ${this.options.getLogFile()} file for more details`,
+          `Api error (102); Check your ${this.options.getLogFile()} file for more details`,
         );
       } else if (code == 103) {
         let error_msg = `Config parsing error (103); Check your ${this.options.getLogFile()} file for more details`;
         if (this.showStatusBar) {
-          this.statusBar.text = '$(clock) WakaTime Error';
-          this.statusBar.tooltip = `WakaTime: ${error_msg}`;
+          this.updateStatusBarText('WakaTime Error');
+          this.updateStatusBarTooltip(`WakaTime: ${error_msg}`);
         }
         this.logger.error(error_msg);
       } else if (code == 104) {
         let error_msg = 'Invalid Api Key (104); Make sure your Api Key is correct!';
         if (this.showStatusBar) {
-          this.statusBar.text = '$(clock) WakaTime Error';
-          this.statusBar.tooltip = `WakaTime: ${error_msg}`;
+          this.updateStatusBarText('WakaTime Error');
+          this.updateStatusBarTooltip(`WakaTime: ${error_msg}`);
         }
         this.logger.error(error_msg);
       } else {
         let error_msg = `Unknown Error (${code}); Check your ${this.options.getLogFile()} file for more details`;
         if (this.showStatusBar) {
-          this.statusBar.text = '$(clock) WakaTime Error';
-          this.statusBar.tooltip = `WakaTime: ${error_msg}`;
+          this.updateStatusBarText('WakaTime Error');
+          this.updateStatusBarTooltip(`WakaTime: ${error_msg}`);
         }
         this.logger.error(error_msg);
       }
     });
   }
 
-  private getCodingActivity(force: boolean = false) {
-    if (!this.showCodingActivity || !this.showStatusBar) return;
+  private getCodingActivity() {
+    if (!this.showStatusBar) {
+      this.clearTodayInterval();
+      return;
+    }
+
+    this.setTodayInterval();
+
+    // prevent updating if we haven't coded since last checked
+    if (this.lastFetchToday > 0 && this.lastFetchToday > this.lastHeartbeat) return;
+
     const cutoff = Date.now() - this.fetchTodayInterval;
-    if (!force && this.lastFetchToday > cutoff) return;
+    if (this.lastFetchToday > cutoff) return;
 
     this.lastFetchToday = Date.now();
-    this.getCodingActivityTimeout = setTimeout(this.getCodingActivity, this.fetchTodayInterval);
 
-    this.hasApiKey(hasApiKey => {
-      if (!hasApiKey) return;
-      this._getCodingActivity();
+    this.options.getApiKey((apiKey) => {
+      if (!apiKey) return;
+      this._getCodingActivity(apiKey);
     });
   }
 
-  private _getCodingActivity() {
+  private _getCodingActivity(apiKey: string) {
     if (!this.dependencies.isCliInstalled()) return;
+
     let user_agent =
       this.agentName + '/' + vscode.version + ' vscode-wakatime/' + this.extension.version;
     let args = ['--today', '--plugin', Utils.quote(user_agent)];
-    if (process.env.WAKATIME_API_KEY) args.push('--key', Utils.quote(process.env.WAKATIME_API_KEY));
+    args.push('--key', Utils.quote(apiKey));
     if (Dependencies.isWindows()) {
       args.push(
         '--config',
-        Utils.quote(this.options.getConfigFile()),
+        Utils.quote(this.options.getConfigFile(false)),
         '--logfile',
         Utils.quote(this.options.getLogFile()),
       );
@@ -507,13 +522,18 @@ export class WakaTime {
     }
     proc.on('close', (code, _signal) => {
       if (code == 0) {
-        if (this.showStatusBar && this.showCodingActivity) {
+        if (this.showStatusBar) {
           if (output && output.trim()) {
-            this.statusBar.text = `$(clock) ${output.trim()}`;
-            this.statusBar.tooltip = 'WakaTime: Today’s coding time. Click to visit dashboard.';
+            if (this.showCodingActivity) {
+              this.updateStatusBarText(output.trim());
+              this.updateStatusBarTooltip('WakaTime: Today’s coding time. Click to visit dashboard.');
+            } else {
+              this.updateStatusBarText();
+              this.updateStatusBarTooltip(output.trim());
+            }
           } else {
-            this.statusBar.text = `$(clock)`;
-            this.statusBar.tooltip = `WakaTime: Calculating time spent today in background...`;
+            this.updateStatusBarText();
+            this.updateStatusBarTooltip('WakaTime: Calculating time spent today in background...');
           }
         }
       } else if (code == 102) {
@@ -549,14 +569,42 @@ export class WakaTime {
   }
 
   private getProjectName(file: string): string {
+    if (!vscode.workspace) return '';
     let uri = vscode.Uri.file(file);
     let workspaceFolder = vscode.workspace.getWorkspaceFolder(uri);
-    if (vscode.workspace && workspaceFolder) {
+    if (workspaceFolder) {
       try {
         return workspaceFolder.name;
       } catch (e) {}
     }
+    if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length) {
+      return vscode.workspace.workspaceFolders[0].name;
+    }
+    return vscode.workspace.name || '';
+  }
+
+  private getProjectFolder(file: string): string {
+    if (!vscode.workspace) return '';
+    let uri = vscode.Uri.file(file);
+    let workspaceFolder = vscode.workspace.getWorkspaceFolder(uri);
+    if (workspaceFolder) {
+      try {
+        return workspaceFolder.uri.fsPath;
+      } catch (e) {}
+    }
+    if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length) {
+      return vscode.workspace.workspaceFolders[0].uri.fsPath;
+    }
     return '';
   }
 
+  private setTodayInterval(): void {
+    if (this.fetchTodayIntervalId) return;
+    this.fetchTodayIntervalId = setInterval(this.getCodingActivity.bind(this), this.fetchTodayInterval);
+  }
+
+  private clearTodayInterval(): void {
+    if (this.fetchTodayIntervalId) clearInterval(this.fetchTodayIntervalId);
+    this.fetchTodayIntervalId = undefined;
+  }
 }